poll works

+++ server/src/streaming/segments/indexes/mod.rs
@@ -3,7 +3,7 @@ mod index_reader;
 mod index_writer;
 
 /// offset: 4 bytes, position: 4 bytes, timestamp: 8 bytes
-pub const INDEX_SIZE: u64 = 16;
+pub const INDEX_SIZE: u32 = 16;
 
 pub use index::Index;
 pub use index::IndexRange;
